Tidy up, style, punctuation
General tidy of the text; yanked a lot of instances of adverbial descriptors ("essentially", etc.): either something's happening or it isn't. 

Applied style guide in a few spots: spelling out numbers under 12 (unless talking about digits or labels: "the rightmost column is 1s", "to represent the number 3", "connected to pin 2" etc.) Particular bugbear: "login" is the noun, "log in" is the verb. Hyphenate adjectival compounds: a four-bit number. Don't hyphenate adverbial compounds: a very strictly controlled diet. Close up words either side of a slash: "on/off", not "on / off". (We had a mixture before.)

Punctuation: added a number of commas for sense. Consistent application of style sheet on bullet punctuation.