Split worksheet into 2
I have made some edits to the text, and have split the worksheet into
two due to its length. I’d suggest that some revisions be made to
section 4 and 5 of worksheet 2, as I think they are moving further from
the main thrust of the resource. A “What Next” section needs to be
added.